% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_fdr_byrun.R
\name{assess_fdr_byrun}
\alias{assess_fdr_byrun}
\title{Calculate the fdr hit rates for every sample in the data.}
\usage{
assess_fdr_byrun(data, FFT = 1, n_range = 20, output = "pdf_csv",
  plot = TRUE, filename = "FDR_report_byrun",
  output_mscore_levels = c(0.01, 0.001))
}
\arguments{
\item{data}{Data structure generated by sample_annotation()}

\item{FFT}{I have read the code for this function a few times now and I still
do not know what this refers to.  I am going to say 'fast fourier
transform', which is obviously not true.  I can say that it is a factor
which is used to change the fdr_cube data, which is very much not the same
thing as an answer.}

\item{n_range}{I am also not certain what this is, nor why 20 is the
optimal default value, but I think the idea is to set up a series of mscore
thresholds.}

\item{output}{Send the result of this to a pdf/csv or the R console?}

\item{plot}{Make some pretty bar plots of the results?}

\item{filename}{Where to send the pdf/csv report (it seems to me this should
be folded into the output parameter).}

\item{output_mscore_levels}{Character list with some predefined thresholds.}
}
\description{
This function calculates some metrics which describe how often each sample
passes the given m score thresholds.  Optionally it may also plot some of
these statistics.
}
