% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert4MSstats.R
\name{convert_for_msstats}
\alias{convert_for_msstats}
\title{Take data from the swath2stats format and get it ready for use by MSstats.}
\usage{
convert_for_msstats(data, replace_values = TRUE, replace_colnames = TRUE,
  replace_unimod = TRUE)
}
\arguments{
\item{data}{SWATH2stats formatted data.}

\item{replace_values}{Change the peculiar sample IDs from openswath into the
somewhat more human-readable format generated from a combination of
run/replicate/etc.}

\item{replace_colnames}{Change a few column names to make them fit better in
msstats?}

\item{replace_unimod}{Change the unimod column name to make it fit better in
msstats?}
}
\value{
MSstats compatible data, hopefully.
}
\description{
Though the two tools use very similar formats, some coercion is required to
pass swath2stats data into msstats; ergo this function.
}
