% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_all_peptides.R
\name{filter_all_peptides}
\alias{filter_all_peptides}
\title{This function is named filter_all_peptides, but it appears to me that it
primarily serves to remove the leading slash number characters from the
peptide names -- which is certainly a good thing(TM), but is that filtering?
Perhaps I am still missing something}
\usage{
filter_all_peptides(data, column = "proteinname", n = 6)
}
\arguments{
\item{data}{SWATH2stats data structure to be modified.}

\item{column}{Which column contains the data to modify?}

\item{n}{How many new IDs should we print to show if this worked?}
}
\value{
Slightly modified data in the same format.
}
\description{
This function is named filter_all_peptides, but it appears to me that it
primarily serves to remove the leading slash number characters from the
peptide names -- which is certainly a good thing(TM), but is that filtering?
Perhaps I am still missing something
}
