% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mscore_condition.R
\name{filter_mscore_condition}
\alias{filter_mscore_condition}
\title{Filter SWATH2stats data based on mscore/condition}
\usage{
filter_mscore_condition(data, mscore = 1, n.replica, rm.decoy = TRUE)
}
\arguments{
\item{data}{SWATH2stats data.}

\item{mscore}{Threshold below which to drop rows.}

\item{n.replica}{Minimum required number of replicates.}

\item{rm.decoy}{Drop decoy rows?}
}
\value{
filtered data structure
}
\description{
Filter SWATH2stats data based on mscore/condition
}
