% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mscore_fdr.R
\name{filter_mscore_fdr}
\alias{filter_mscore_fdr}
\title{Filter data based on a given FDR and mscore. (I think)}
\usage{
filter_mscore_fdr(data, FFT = 1, overall_protein_fdr_target = 0.02,
  mscore_limit = 0.01, upper_overall_peptide_fdr_limit = 0.05,
  rm.decoy = TRUE)
}
\arguments{
\item{data}{SWATH2stats data structure to filter.}

\item{FFT}{Look, I know this is not actually 'fast fourier transform', but
that is still my best guess.}

\item{overall_protein_fdr_target}{Choose an fdr target to hit.}

\item{mscore_limit}{and an mscore threshold.}

\item{upper_overall_peptide_fdr_limit}{phew, that is a mouthful, but it is
what it says on the tin.}

\item{rm.decoy}{Drop decoy rows?}
}
\value{
filtered data!
}
\description{
Filter data based on a given FDR and mscore. (I think)
}
