% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mscore_freqobs.R
\name{filter_mscore_freqobs}
\alias{filter_mscore_freqobs}
\title{Filter openSWATH output table according to mscore.}
\usage{
filter_mscore_freqobs(data, mscore = 0.01, percentage = NULL,
  rm.decoy = TRUE)
}
\arguments{
\item{data}{A data frame containing SWATH data.}

\item{mscore}{Value that defines the mscore threshold according to which the
data will be filtered.}

\item{percentage}{Percentage in which replicas the transition has to reach
the mscore threshold.}

\item{rm.decoy}{Option to remove the decoys during filtering.}
}
\value{
Returns a data frame with the filtered data.
}
\description{
This function filters the SWATH data according to the m_score value, as well
as to the number of occurence in the data (requant) and within a condition
(condition)
}
\examples{
\dontrun{
 data("OpenSWATH_data", package="SWATH2stats")
 data("Study_design", package="SWATH2stats")
 data <- sample_annotation(OpenSWATH_data, Study_design)
 data.filtered <- filter_mscore(data, 0.01)
 data.filtered <- filter_mscore_freqobs(data, 0.01, 0.8)
 data.filtered <- filter_mscore_condition(data, 0.01, 3)
}
}
\author{
Peter Blattmann
}
