% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_on_max_peptides.R
\name{filter_on_max_peptides}
\alias{filter_on_max_peptides}
\title{Filter-out absurdly high peptides from the data.}
\usage{
filter_on_max_peptides(data, n_peptides = 6, rm.decoy = TRUE,
  column = "proteinname")
}
\arguments{
\item{data}{SWATH2stats data to filter.}

\item{n_peptides}{Number of first n peptides to keep}

\item{rm.decoy}{Drop the decoys?}

\item{column}{which column to use for filtering?}
}
\value{
filtered data!
}
\description{
Filter-out absurdly high peptides from the data.
}
