% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_proteotypic_peptides.R
\name{filter_proteotypic_peptides}
\alias{filter_proteotypic_peptides}
\title{Filter data to keep rows which make sense with respect to proteolytic ends.}
\usage{
filter_proteotypic_peptides(data, rm.decoy = TRUE, column = "proteinname")
}
\arguments{
\item{data}{SWATH2stats data to filter.}

\item{rm.decoy}{Drop decoys?}

\item{column}{Which column to query for filtering?}
}
\value{
Filtered data!
}
\description{
Filter data to keep rows which make sense with respect to proteolytic ends.
}
