% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JPP_update.R
\name{JPP_update}
\alias{JPP_update}
\title{Select alternate m_score column in JPP data and avert user.}
\usage{
JPP_update(data, mscore.col)
}
\arguments{
\item{data}{Data table produced by the OpenSWATH/pyProphet workflow.}

\item{mscore.col}{The column from which to retrieve the m_score data.  If one uses JPP
(Rosenberger, Bludau et al. 2017) this can be used to select between Protein
and transition_group m_score.}
}
\value{
The mscore column that may have been changed to
  transition_group_id_m_score.  Also this returns a message to the user.
}
\description{
Thr output from JPP (Rosenberge, Bludau et al. 2017) no longer has the
m_score column, but a ProteinName_m_score and
transition_group_id_m_score.  To make users aware of this, this function
tests if the m_score still exists and selects as default the
transition_group_id_m_score column.
}
\examples{
  data("OpenSWATH_data", package="SWATH2stats")
  JPP_update(OpenSWATH_data, "m_score")
}
\author{
Peter Blattmann
}
