% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mscore4assayfdr.R
\name{mscore4assayfdr}
\alias{mscore4assayfdr}
\title{Calculate the mscore which provides the given assay fdr hit rate.}
\usage{
mscore4assayfdr(data, FFT = 1, fdr_target = 0.01)
}
\arguments{
\item{data}{SWATH2stats data to poke.}

\item{FFT}{I dunno.}

\item{fdr_target}{What fdr to query?}
}
\value{
an mscore which gets to this fdr hit rate.
}
\description{
Calculate the mscore which provides the given assay fdr hit rate.
}
