% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mscore4protfdr.R
\name{mscore4protfdr}
\alias{mscore4protfdr}
\title{Calculate the mscore which provides the given protein fdr hit rate.}
\usage{
mscore4protfdr(data, FFT = 1, fdr_target = 0.02)
}
\arguments{
\item{data}{SWATH2stats data to poke.}

\item{FFT}{I dunno.}

\item{fdr_target}{What fdr to query?}
}
\value{
an mscore which gets to this fdr hit rate.
}
\description{
Hmmm the 3 functions in this family are largely equivalent, differing only in
the column which is used to query the name.  At some point in the near future
I would like to replace these with a generic function that can handle either
assays, peptides, or proteins.
}
