% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSstats_data}
\alias{MSstats_data}
\title{Testing dataset in MSstats format.}
\description{
A small table with the column names corresponding to the MSstats format. 
 This data is intended only to test functions.
}
\author{
Peter Blattmann
}
