% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OpenSWATH_data}
\alias{OpenSWATH_data}
\alias{data}
\title{Testing dataset from OpenSWATH.}
\description{
A small selection of the data obtained from the iPortal pipeline for an 
 SWATH/DIA experiment with perturbations relating to cholesterol regulation. 
 Protein and Peptides have been anonymized as the data is unpublished.
 The FDR version of the test data contains modified (lowered) decoy peak 
 group m_scores to simulate FDR behaviour of a large dataset.
}
\author{
Peter Blattmann
}
