% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation_between_samples.R
\name{plot_correlation_between_samples}
\alias{plot_correlation_between_samples}
\title{Make a plot of the correlation between samples/groups in the swath2stats data.}
\usage{
plot_correlation_between_samples(data, column.values = "intensity",
  size = 6, comparison = transition_group_id ~ condition + bioreplicate,
  fun.aggregate = NULL, label = TRUE, ...)
}
\arguments{
\item{data}{SWATH2stats data to poke.}

\item{column.values}{Which column to query?}

\item{size}{How large should the text in the grid be (smaller is better for
rmarkdown html reports).}

\item{comparison}{Formula describing the organization of samples to query.}

\item{fun.aggregate}{Which function should be used to gather the data?}

\item{label}{Put labels of the correlations in the plot?}
}
\value{
plot!
}
\description{
Make a plot of the correlation between samples/groups in the swath2stats data.
}
