% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fdr_table.R
\name{plot_fdr_table}
\alias{plot_fdr_table}
\title{Plot fdr metrics for multiple mscore levels.}
\usage{
plot_fdr_table(x, output = "Rconsole", filename = "FDR_report_overall", ...)
}
\arguments{
\item{x}{Data structure generated by sample_annotation()}

\item{output}{Send the results to a pdf or the R console?}

\item{filename}{When printing to a file, what file?}

\item{...}{Extra arguments passed on to functions inside this.}
}
\value{
Originally this returned nothing, but now it makes a list of ggplot2
  plots which may be passed along and plotted as desired (with that in mind,
  I would like to remove the explicit plot() calls in this function).
}
\description{
This function provides a set of plots for the assays, peptides, and proteins
comprising the number of observations and the mscore hit-rate for each mscore
threshold provided.
}
