% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variation_vs_total.R
\name{plot_variation_vs_total}
\alias{plot_variation_vs_total}
\title{Plot variation data!}
\usage{
plot_variation_vs_total(data, column.values = "intensity",
  comparison1 = transition_group_id ~ bioreplicate + condition,
  comparison2 = transition_group_id + condition ~ bioreplicate,
  fun.aggregate = NULL, label = TRUE, ...)
}
\arguments{
\item{data}{SWATH2stats data to poke.}

\item{column.values}{Which column to poke?}

\item{comparison1}{A formula describing the group(s) to examine.}

\item{comparison2}{Another formula describing the group(s) to examine.}

\item{fun.aggregate}{Which function should be used to bring together the
values (I think this should be sum, yesno?)}

\item{label}{Put labels in the plot?}
}
\value{
a plot!
}
\description{
I have not looked closely at this function yet.
}
