% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeDecoyProteins.R
\name{removeDecoyProteins}
\alias{removeDecoyProteins}
\title{Use a couple of small regular expressions to drop decoy rows.}
\usage{
removeDecoyProteins(data, column = "proteinname")
}
\arguments{
\item{data}{SWATH2stats data to filter.}

\item{column}{which column to query?}
}
\value{
filtered data.
}
\description{
Use a couple of small regular expressions to drop decoy rows.
}
