% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeDecoyProteins.R
\name{rmDecoyProt}
\alias{rmDecoyProt}
\title{What string defines decoys?}
\usage{
rmDecoyProt(x, pattern = "DECOY")
}
\arguments{
\item{x}{proteinname string to query.}

\item{pattern}{chosen string to seek out.}
}
\description{
What string defines decoys?
}
