% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_annotation.R
\name{sample_annotation}
\alias{sample_annotation}
\title{Given dataframes of TRIC processed data and sample annotations, mash them
together into something appropriate for downstream analyses.}
\usage{
sample_annotation(data, sample_annotation, data_type = "OpenSWATH",
  annotation_file_column = "filename", data_file_column = "filename",
  condition_column = "condition", replicate_column = "bioreplicate",
  fullpeptidename_column = "fullunimodpeptidename", run_id = NULL,
  run_column = "run", change_run_id = TRUE, verbose = FALSE)
}
\arguments{
\item{data}{Data frame following processing by feature_alignment.py.}

\item{data_type}{Either OpenSWATH or MSstats, this defines the columns which
get created in the output data frame.}

\item{annotation_file_column}{Name of the column containing the output file
from the experiment metadata for  each sample.  In my own sample sheet, I keep
columns for the mzXML files, tric outputs, raw files, and osw outputs from
OpenSwathWorkFlow; and I cannot be relied upon to remember which is which,
ergo this option.}

\item{data_file_column}{Column in the feature_alignment.py containing the
mzXML file for each detected transition.}

\item{condition_column}{Which column annotates the experimental condition in
the swath data?}

\item{replicate_column}{Which column annotates the replicate in the data?}

\item{run_column}{Which column annotates the separate runs?}

\item{change_run_id}{Concatenate the condition, replicate, and run columns
into a new run_id column?}

\item{verbose}{Be verbose?}
}
\value{
Dataframe with each row annotated from the sample metadata.
}
\description{
This performs some quick sanity checks on the data and annotations and
creates the 'Condition', 'BioReplicate', and 'Run' columns along with other
columns expected by MSstats/OpenSWATH.
}
\examples{
 \dontrun{
   data("OpenSWATH_data", package="SWATH2stats")
   data("Study_design", package="SWATH2stats")
   data <- sample_annotation(OpenSWATH_data, Study_design)
}
}
\author{
Peter Blattman
}
