% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Spyogenes}
\alias{Spyogenes}
\title{S.pyogenes example data.}
\description{
A table containing SWATH-MS data from S.pyogenes.
 This table was generated from the original data deposited on PeptideAtlas 
 (PASS00289, file "rawOpenSwathResults_1pcnt_only.tsv") by selecting only 
 the column necessary for the SWATH2stats.
}
\references{
Rost, H. L., et al. (2014). OpenSWATH enables automated, 
  targeted analysis of data-independent acquisition MS data. 
  Nat Biotechnol 32(3): 219-223.
}
