% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Study_design}
\alias{Study_design}
\title{A set of sample annotations which describe an experiment 6 of HeLa control
and treated samples (3 of each).}
\format{a data frame of 6 rows and 4 columns.}
\usage{
data(Study_design)
}
\description{
This provides Filename, Condition, BioReplicate, and Run data for a small
example dataset.  I assumed at first that these correspond to the data in
MSstats_data, but that cannot be, as they have different Run IDs.
}
\examples{
 data(Study_design)
 dim(Study_design)
}
\keyword{datasets}
