\name{SWATH2stats-package}
\alias{SWATH2stats-package}
\alias{SWATH2stats}
\docType{package}
\title{
SWATH2stats
}
\description{
This package is intended to transform SWATH data from the OpenSWATH
    software into a format readable by other statistics packages while performing
    filtering, annotation and FDR assessment.
}
\details{
\tabular{ll}{
Package: \tab SWATH2stats\cr
Type: \tab Package\cr
Version: \tab 1.3.6\cr
Date: \tab 2016-06-28\cr
License: \tab GPLv3\cr
}
}
\author{
Peter Blattmann, Moritz Heusel and Ruedi Aebersold

Maintainer: Peter Blattmann <blattmann@imsb.biol.ethz.ch>
}
\references{
Blattmann P, Heusel M, Aebersold R. SWATH2stats: An R/Bioconductor Package to Process and Convert
Quantitative SWATH-MS Proteomics Data for Downstream Analysis Tools. PLoS ONE 11(4): e0153160 (2016). doi: 10.1371/journal.pone.0153160.

Rost HL, Rosenberger G, Navarro P, Gillet L, Miladinovic SM, Schubert OT, Wolski W, Collins BC, Malmstrom J, Malmstrom L, Aebersold R. OpenSWATH enables automated, targeted analysis of data-independent acquisition MS data. Nature Biotechnology. 2014 Mar;32(3):219-23. doi: 10.1038/nbt.2841.

Choi M, Chang CY, Clough T, Broudy D, Killeen T, MacLean B, Vitek O. MSstats: an R package for statistical analysis of quantitative mass spectrometry-based proteomic experiments.Bioinformatics. 2014 Sep 1;30(17):2524-6. doi: 10.1093/bioinformatics/btu305.

Rosenberger G, Ludwig C, Rost HL, Aebersold R, Malmstrom L. aLFQ: an R-package for estimating absolute protein quantities from label-free LC-MS/MS proteomics data. Bioinformatics. 2014 Sep 1;30(17):2511-3. doi: 10.1093/bioinformatics/btu200.
}
\keyword{SWATH2stats}
\seealso{
\code{\link[aLFQ]{aLFQ}},
\code{\link[MSstats]{MSstats}},
}
