% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWATH2stats.R
\docType{package}
\name{SWATH2stats}
\alias{SWATH2stats}
\alias{SWATH2stats-package}
\title{SWATH2stats: a suite of tools to filter, plot, and convert DIA data.}
\description{
SWATH2stats: a suite of tools to filter, plot, and convert DIA data.
}
