% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_MSstats_OpenSWATH.R
\name{transform_MSstats_OpenSWATH}
\alias{transform_MSstats_OpenSWATH}
\title{I have no idea what this function does, lets look at it...}
\usage{
transform_MSstats_OpenSWATH(data)
}
\arguments{
\item{data}{swath2stats data to modify, I am guessing.}
}
\value{
transformed data?
}
\description{
I have no idea what this function does, lets look at it...
}
