% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifyProteinGroupLabels.R
\name{unifyProteinGroupLabels}
\alias{unifyProteinGroupLabels}
\title{Bring together protein group labels.}
\usage{
unifyProteinGroupLabels(data, column = "proteinname")
}
\arguments{
\item{data}{SWATH2stats data to modify.}

\item{column}{Which column to use for unifying the groups.}
}
\value{
slightly modified data structure.
}
\description{
Bring together protein group labels.
}
