% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_matrix_proteins.R
\name{write_matrix_proteins}
\alias{write_matrix_proteins}
\title{Write out a matrix of proteins from swath2stats data.}
\usage{
write_matrix_proteins(data, write.csv = FALSE, fun.aggregate = sum,
  filename = "SWATH2stats_overview_matrix_proteinlevel.csv",
  rm.decoy = FALSE)
}
\arguments{
\item{data}{SWATH2stats data to write out.}

\item{write.csv}{Write the result to a csv file?  This should be folded into the next arg.}

\item{fun.aggregate}{What function to use when aggregating the set of intensities?}

\item{filename}{Where to write the data?}

\item{rm.decoy}{Remove the decoys?}
}
\value{
the peptides as a matrix!
}
\description{
This function should be folded into the write_matrix_peptides.
}
