% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_protein_ids.R
\name{add_genesymbol}
\alias{add_genesymbol}
\title{Gather gene symbols from biomart and add them to a swath2stats data set.}
\usage{
add_genesymbol(data_table, gene.ID.table, column.name = "protein",
  ID1 = "uniprotswissprot", ID2 = "hgnc_symbol", id.separator = "/",
  copy_nonconverted = TRUE)
}
\arguments{
\item{data_table}{A data frame or file name.}

\item{gene.ID.table}{a table to match against}

\item{column.name}{The column name where the original protein identifiers
are present.}

\item{ID1}{The type of the original protein identifiers
(e.g. "uniprotswissprot", "ensembl_peptide_id").}

\item{ID2}{The type of the converted protein identifiers
(e.g. "hgnc_symbol", "mgi_symbol", "external_gene_name").}

\item{id.separator}{Separator between protein identifiers of shared
peptides.}

\item{copy_nonconverted}{Option defining if the identifiers that cannot be
converted should be copied.}
}
\value{
some gene symbols
}
\description{
Gather gene symbols from biomart and add them to a swath2stats data set.
}
\note{
Protein identifiers from shared peptides should be separated by a forward
slash. The host of archived ensembl databases can be introduced as well
(e.g. "dec2017.archive.ensembl.org")
}
\author{
Peter Blattmann
}
