% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_decoy_rate.R
\name{assess_decoy_rate}
\alias{assess_decoy_rate}
\title{Query a swath2stats data set for decoy hit rates.}
\usage{
assess_decoy_rate(data)
}
\arguments{
\item{data}{A data frame that contains at least a column named "FullPeptideName" and "decoy".}
}
\value{
list containing the number of decoys, non-decoys, and the ratio.
}
\description{
This looks at the swath2stats table and compares the length of the detected
decoy peptides vs non-decoy peptides.
}
\details{
A printout is generated to indicate the number of non-decoy, decoy peptides
and the rate of decoy vs non-decoy peptides. Unique peptides are counted, so
a precursor with different charge states is counted as one peptide. In the
column "decoy" the values need to be 1,0 or TRUE and FALSE.
}
\examples{
\dontrun{
 data("OpenSWATH_data", package="SWATH2stats")
 data <- OpenSWATH_data
 assess_decoy_rate(data)
}
}
\author{
Peter Blattmann
}
