% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_decoy_rate.R
\name{assess_decoy_rate}
\alias{assess_decoy_rate}
\title{Query a swath2stats data set for decoy hit rates.}
\usage{
assess_decoy_rate(data)
}
\arguments{
\item{data}{SWATH2stats data structure.}
}
\value{
list containing the number of decoys, non-decoys, and the ratio.
}
\description{
This looks at the swath2stats table and compares the length of the detected
decoy peptides vs non-decoy peptides.
}
