% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_analytes.R
\name{count_analytes}
\alias{count_analytes}
\title{Count the number of analytes observed in a swath2stats data structure.}
\usage{
count_analytes(data, column_levels = c("transition_group_id",
  "fullpeptidename", "proteinname"), column_by = "run_id", rm_decoy = TRUE)
}
\arguments{
\item{data}{swath2stats data structure.}

\item{column_levels}{Use these columns to collate the data with stats::aggregate.}

\item{column_by}{Use this column for collecting the resulting counts.}

\item{rm_decoy}{Keep the decoys?}
}
\value{
data frame of counted peptides.
}
\description{
Count the number of analytes observed in a swath2stats data structure.
}
