% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregate.R
\name{disaggregate}
\alias{disaggregate}
\title{I do not yet understand what is going on in this function.  It is querying
the swath2stats data for a few parameters and eventually melting
it. Depending on what comes of that, it drops some rows and returns a new,
presumably better data structure.}
\usage{
disaggregate(data, all.columns = FALSE)
}
\arguments{
\item{data}{SWATH2stats data structure to beat on.}

\item{all.columns}{Mess with the set of returned columns?}
}
\value{
a new version of the swath2stats data.
}
\description{
I do not yet understand what is going on in this function.  It is querying
the swath2stats data for a few parameters and eventually melting
it. Depending on what comes of that, it drops some rows and returns a new,
presumably better data structure.
}
