% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mscore.R
\name{filter_mscore}
\alias{filter_mscore}
\title{Drop rows in the SWATH2stats data which have mscores less than or equal to
the given mscore threshold.}
\usage{
filter_mscore(data, mscore = 1, rm.decoy = TRUE)
}
\arguments{
\item{data}{SWATH2stats data structure.}

\item{mscore}{Threshold below-which data is unwanted.}

\item{rm.decoy}{Drop decoys from the data?}
}
\value{
A copy of the data which has been mscore-filtered.
}
\description{
Drop rows in the SWATH2stats data which have mscores less than or equal to
the given mscore threshold.
}
