% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mscore_freqobs.R
\name{filter_mscore_freqobs}
\alias{filter_mscore_freqobs}
\title{Filter data based on a given FDR and mscore. (I think)}
\usage{
filter_mscore_freqobs(data, mscore = 0.01, percentage = NULL,
  rm.decoy = TRUE)
}
\arguments{
\item{data}{SWATH2stats data structure to filter.}

\item{mscore}{Chosen mscore limit.}

\item{percentage}{Look for peptides in > this percent of the possible
conditions, if they are found in less than this threshold, bye bye!}

\item{rm.decoy}{Drop the decoys!?}
}
\value{
filtered data!
}
\description{
Filter data based on a given FDR and mscore. (I think)
}
