% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import some data into swath2stats, presumably from openMS.}
\usage{
import_data(data)
}
\arguments{
\item{data}{I am guessing tric-formatted tsv data?}
}
\value{
swath2stats data
}
\description{
Import some data into swath2stats, presumably from openMS.
}
