\name{MSstats_data}
\alias{MSstats_data}
\docType{data}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Testing dataset in MSstats format
}
\description{
A small table with the column names corresponding to the MSstats format. This data is intended only to test functions. 
}
\author{
Peter Blattmann
}
