% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSstats_data}
\alias{MSstats_data}
\title{6 rows of sample MSstats data.}
\format{a data frame of 6 rows and 11 columns.}
\usage{
data(MSstats_data)
}
\description{
This provides 6 intensities of putative MSstats data for 3 replicates of HeLa cells.
}
\examples{
 data(MSstats_data)
 dim(MSstats_data)
}
\keyword{datasets}
