% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OpenSWATH_data}
\alias{OpenSWATH_data}
\title{Sample data from the OpenMS, pyprophet workflow.}
\format{a data frame of 2369 rows and 58 columns.}
\usage{
data(OpenSWATH_data)
}
\description{
I am not sure where this data came from, I need to look a little more closely.
}
\examples{
 data(OpenSWATH_data)
 dim(OpenSWATH_data)
}
\keyword{datasets}
