% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_OpenSWATH_output.R
\name{reduce_OpenSWATH_output}
\alias{reduce_OpenSWATH_output}
\title{OpenSwathWorkflow provides a bunch of columns which are not required for mapDIA and MSstats.
This function will drop those extraneous columns.}
\usage{
reduce_OpenSWATH_output(data, column.names = NULL)
}
\arguments{
\item{data}{SWATH2stats data to modify.}

\item{column.names}{Set of columns to seek out.}
}
\value{
filtered data.
}
\description{
OpenSwathWorkflow provides a bunch of columns which are not required for mapDIA and MSstats.
This function will drop those extraneous columns.
}
