% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeDecoyProteins.R
\name{rmDecoyProt}
\alias{rmDecoyProt}
\title{Subfunction to remove decoys}
\usage{
rmDecoyProt(x, decoy_string = "DECOY")
}
\arguments{
\item{x}{proteinname string to query.}

\item{decoy_string}{String defining a decoy}
}
\value{
returns string without elements containing the decoy string
}
\description{
Subfunction to remove decoys
}
