% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_annotation.R
\name{sample_annotation}
\alias{sample_annotation}
\title{Annotate the SWATH data with the sample information}
\usage{
sample_annotation(data, sample_annotation, data_type = "OpenSWATH",
  annotation_file_column = "filename", data_file_column = "filename",
  condition_column = "condition", replicate_column = "bioreplicate",
  fullpeptidename_column = "fullunimodpeptidename", run_id = NULL,
  run_column = "run", change_run_id = TRUE, verbose = FALSE)
}
\arguments{
\item{data}{A data frame containing SWATH data.}

\item{sample_annotation}{A data frame containing the columns: Filename,
Condition, BioReplicate, Run. The values contained in the column filename
have to be present in the filename of the SWATH data.}

\item{data_type}{Option to specify the format of the table, if the column
names from an OpenSWATH output or MSstats table are used.}

\item{annotation_file_column}{Name of the column containing the output file
from the experiment metadata for  each sample.  In my own sample sheet, I keep
columns for the mzXML files, tric outputs, raw files, and osw outputs from
OpenSwathWorkFlow; and I cannot be relied upon to remember which is which,
ergo this option.}

\item{data_file_column}{Option to specify the column name where the injection
file is specified. Default is set to "filename".}

\item{condition_column}{Which column annotates the experimental condition in
the swath data?}

\item{replicate_column}{Which column annotates the replicate in the data?}

\item{run_column}{Which column annotates the separate runs?}

\item{change_run_id}{Option to choose if the run\_id column shall be
reassigned to a unique value combining the values of Condition,
BioReplicate and Run. (Option only possible if data is of format
"OpenSWATH")}

\item{verbose}{Option to turn on reporting on which filename it is working
on.}
}
\value{
Returns a dataframe with each row annotated for the study design
}
\description{
For statistical analysis and filtering the measurements need to be annotated
with Filename, Condition, BioReplicate, and Run. This functions takes this
information from a txt file containing this meta-data.
}
\details{
Given dataframes of TRIC processed data and sample annotations, mash them
together into something appropriate for downstream analyses.

This performs some quick sanity checks on the data and annotations and
creates the 'Condition', 'BioReplicate', and 'Run' columns along with other
columns expected by MSstats/OpenSWATH.
}
\examples{
 \dontrun{
   data("OpenSWATH_data", package="SWATH2stats")
   data("Study_design", package="SWATH2stats")
   data <- sample_annotation(OpenSWATH_data, Study_design)
}
}
\author{
Peter Blattman
}
