% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Spyogenes}
\alias{Spyogenes}
\title{A set of Streptococcus pyogenes intensities from 4 samples.}
\format{a data frame of 32,272 rows and 13 columns.}
\usage{
data(Spyogenes)
}
\description{
To my eyes, this appears to be the result of OpenMS->pyprophet with the
exception that the alignment files look to be gzipped Excel files.  I bet
they are actually tab separated outputs from feature_alignment.py or
pyprophet.
}
\examples{
 data(Spyogenes)
 data <- Spyogenes
 dim(data)
}
\keyword{datasets}
