\docType{data}
\name{Study_design}
\alias{Study_design}

\title{
Study design table
}
\description{
A table containing the meta-data defining the study design.

\describe{
\item{Filename}{
A unique identifier corresponding to the filename in the SWATH data.
}
\item{Condition}{
The Condition explains the perturbation performed on this sample.
}
\item{BioReplicate}{
Number indicating the biological replicate of this sample. 
}
\item{Run}{
A unique number for each MS-injection. 
}
}
}

\source{Peter Blattmann}
\author{Peter Blattmann}
