% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Study_design}
\alias{Study_design}
\title{A table containing the meta-data defining the study design of the OpenSWATH data.}
\description{
This table contains a unique identifier in the column "Filename"
 corresponding to the filename in the SWATH data.
 In the column "Condition" the perturbation performed is described.
 In the column "BioReplicate" the biological replicate is indicated.
 In the column "Run" a unique identifier for each injection.
 Technical injections would have different Run numbers but the same 
 BioReplicate number.
}
\author{
Peter Blattmann
}
